upgrade to new directory structure
